pyproject.toml
[project]
name = "gepa_artifact"
version = "0.0.1"
description = "GEPA is a novel prompt optimizers that uses text feedback to improve compound AI systems."
authors = [
  { name = "Lakshya A Agrawal", email = "lakshyaaagrawal@berkeley.edu" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bm25s==0.2.12",
    "dspy",
    "fasteners==0.19",
    "func-timeout==4.3.5",
    "jax[cpu]==0.6.0",
    "matplotlib>=3.10.3",
    "pystemmer==2.2.0.3",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.13",
    "tabulate>=0.9.0",
    "pylcs==0.1.1",
    "seaborn",
    "adjustText",
    "diskcache",
    "evaluate==0.4.3",
    "fqdn==1.5.1",
    "httptools==0.6.4",
    "isoduration==20.11.0",
    "jupyter==1.1.1",
    "langchain-chroma==0.2.3",
    "langchain-community==0.3.23",
    "langchain-huggingface==0.1.2",
    "psutil==7.0.0",
    "sacrebleu==2.5.1",
    "tinycss2==1.4.0",
    "uri-template==1.3.0",
    "uvloop==0.21.0",
    "wandb==0.19.10",
    "watchfiles==1.0.5",
    "webcolors==24.11.1",
    "websockets==15.0.1",
    "absl-py",
    "langdetect",
    "nltk",
    "immutabledict",
    "spacy",
    "emoji",
    "syllapy",
    "unicodedata2",
    "pip",
    "math-verify==0.6.0",
    "lark",
    "sympy",
    "levenshtein",
    "aicodetools>=0.1.7",
]

[tool.uv.sources]
arbor-ai = { path = "gepa_artifact/utils/arbor", editable = true }
dspy = { path = "gepa_artifact/utils/dspy", editable = true }

[tool.uv]
no-build-isolation-package = ["flash-attn"]

[[tool.uv.dependency-metadata]]
name = "flash-attn"
version = "2.7.4.post1"
requires-dist = ["torch", "setuptools"]
